split code inside api
