+++ documentsearcher.vcxproj
@@ -21,13 +21,14 @@
   <ItemGroup>
     <ClCompile Include="src\documentsearcher.cpp" />
     <ClCompile Include="src\logger.cpp" />
+    <ClCompile Include="src\lucene_api\api.cpp" />
     <ClCompile Include="src\lucene_api\index.cpp" />
     <ClCompile Include="src\lucene_api\search.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\logger.h" />
     <ClInclude Include="src\lucene_api\index.h" />
-    <ClInclude Include="src\lucene_api\lucene_api.h" />
+    <ClInclude Include="src\lucene_api\api.h" />
     <ClInclude Include="src\lucene_api\search.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
