+++ src/lucene_api/api.cpp
@@ -1,0 +1,26 @@
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
+
+#ifndef NOMINMAX
+#define NOMINMAX
+#endif
+
+#include "api.h"
+
+#include <codecvt>
+#include <iostream>
+
+namespace LuceneAPI {
+    std::wstring utf8ToUtf16(const std::string& utf8Str)
+    {
+        std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>> conv;
+        return conv.from_bytes(utf8Str);
+    }
+
+    std::string utf16ToUtf8(const std::wstring& utf16Str)
+    {
+        std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>> conv;
+        return conv.to_bytes(utf16Str);
+    }
+}
